requirements.txt
flask==3.0.3
flask-cors==4.0.1
Pillow==10.4.0
imagehash==4.3.1
scikit-image==0.24.0
numpy==1.26.4
gunicorn==22.0.0
